Stage 4 LUT trained and debugged — ΔE2000 45→24.7, MS-SSIM 0.32→0.55
Three training runs, three bugs found:

Run 1 (C1 TIFs as input): domain mismatch — trained on C1 (bright=117) but
inferred on our S3 output (bright=80). LUT barely helped.

Run 2 (our S1+S3 as input): degenerate init — all n_luts initialised as
identical identity → classifier weights collapse to [0,0,1], only one LUT
trains. Brightening factor 1.097× instead of needed ~2×.

Run 3 (diverse init + brightness norm):
- Basis LUTs init with distinct roles: identity / gamma-brightened / warm grade
- Brightness normalisation before LUT decouples per-image exposure from colour
  grade, so LUT learns only the aesthetic transform
- LUT diversity: 0.068 inter-LUT diff (vs 0.00009 before)
- Best loss 0.041971 → ΔE2000 24.70, MS-SSIM 0.5530 vs Matt reference

Remaining gap vs C1 (ΔE=14.23): R/B cast 0.905 vs target 1.049, and
brightness 135.6 vs Matt's 159.3. Root cause: 21 pairs + no GPU + highlight
clipping in the linear brightness scale. More pairs or LPIPS loss will close.

+++ src/stage1_raw/lens_correction.py
@@ -73,7 +73,7 @@ def correct_lens(
 
     log.info("lens_correction.ok", body=exif.get("camera_body"), lens=str(lens))
 
-    result = np.clip(work * scale, 0, scale).astype(original_dtype)
+    result = np.clip(work.astype(np.float64) * scale, 0, scale).astype(original_dtype)
     return result
